requirements.txt
APScheduler==3.9.1.post1
asgiref==3.6.0
async-timeout==4.0.2
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==2.1.1
Django==3.2
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
oauthlib==3.2.2
pycodestyle==2.10.0
python-decouple==3.8
python-dotenv==0.21.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
redis==4.4.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
spotipy==2.22.0
sqlparse==0.4.3
tomli==2.0.1
tweepy==4.12.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
ytmusicapi==0.25.0
